pyproject.toml
[project]
name = "shazam"
version = "0.1.0"
description = "trying to make shazam in python"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "requests",
    "yt-dlp",
    "ffmpeg-python",
    "fastapi[standard]",
    "torch",
    "numpy<2",
    "torchaudio",
    "ipykernel",
    "librosa==0.11",
    "numba==0.61.2",
    "llvmlite==0.44.0",
    "matplotlib>=3.10.8",
    "opencv-python",
    "ruff>=0.14.13",
]
[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'"
]
